packages/ai-agents/requirements.txt
openai==1.12.0
pydantic==2.6.1
python-dotenv==1.0.0
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.5
web3==6.11.4
eth-typing==3.5.2
eth-utils==2.3.1